Funções movidas para um módulo para serem aproveitadas
